Refactor src into packages
- Enforce clean-arch
