+++ packages/application/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@fp-ts-playground/application",
+  "version": "1.0.0",
+  "description": "Application package",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "build": "tsc"
+  },
+  "repository": "https://github.com/danielmunafo/fp-ts-playground.git",
+  "author": "dann.tostes@gmail.com",
+  "license": "MIT",
+  "private": true,
+  "dependencies": {
+    "@fp-ts-playground/core": "*",
+    "@fp-ts-playground/infrastructure": "*"
+  }
+}
