+++ packages/application/src/use-case/errors.ts
