+++ packages/application/src/use-case/registerUser.ts
@@ -1,7 +1,7 @@
+import { User, UserRepositoryPort } from "@fp-ts-playground/core";
 import * as O from "fp-ts/Option";
 import * as TE from "fp-ts/TaskEither";
 import { pipe } from "fp-ts/function";
-import { User, UserRepositoryPort } from "../../core";
 import { EmailInUseError } from "./errors";
 
 export const registerUser =
