+++ packages/core/src/domain/entities/UserEntity.ts
