+++ packages/core/src/domain/entities/index.ts
@@ -1,1 +1,2 @@
+export * from "./UserEntity";
 export * from "./UserEntityArb";
