+++ packages/core/src/domain/index.ts
